Time: 16 ms (20.28%) | Memory: 35 MB (88.76%) - LeetSync

+++ 1056-capacity-to-ship-packages-within-d-days/capacity-to-ship-packages-within-d-days.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    bool possible(vector<int>&a,long long  guess,int days)
+    bool possible(vector<int>&a,int   guess,int days)
     {
         int day=1;
         int load = 0 ;
